Perfect Number Generator
A program where you input how many perfect numbers you want and it calculates them and prints them out. The method of to calculate a perfect number is: ((2^p) - 1)(2^(p-1)).  This is only a perfect number if the Mersenne Number is prime ((2^p)-1). For it to be prime p must be a prime number, hence the prime number generator embedded in the code. To check that the Mersenne Number is prime the program uses a Lucas–Lehmer primality test. This is how it finds the next prime Mersenne  Number and therefore perfect number. :)